Added a few more examples to the examples directory of Proboscis. These examples are designed to be run the same way an end-user might setup and run Proboscis (versus the Nose tests).
Changed the build_sphinx configuration to actually execute these examples and save the output to later display as part of the documentation. This means the output seen in the example docs is real and up to date as of the time Sphinx generated the doc.

Added authentic TestNG style test methods, and attempted to emulate the same behavior of instances. Now if the @test decorator is found on class methods as well as the class itself, the methods aren't treated as functions but as real methods that all run with the same instantiated instance.  The "depends_on" property works with these methods, meaning the "dependsOnMethod" functionality from TestNG works here. Additionally a method can still depend on a Class, and tests can depend on these new test methods by depending on their parent class.

I'm not noticing speed issues with sorting yet but in a vain effort to mitigate them I changed some things from lists to sets.

Created @before_class and @after_class which work exactly like @test, but run before all other methods as well as after all other methods respectively.

Created @factory which tries to emulate the @factory decorator for TestNG.  However this only works on free-standing functions, and is executed at start-up before any tests are run. The tests it generates are sorted though. TestNG is much more dynamic that it even initially seems so I wonder how far I can get sorting everything initially and passing it to Nose...

I created a module for assertions similar to TestNG's. Nose has some like this in its tools module but I figured there might still be a use for these, especially since in the future maybe Nose could become optional instead of required. Additionally some people might appreciate the flipped ordering.